Tourist-api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
boto3==1.37.11
botocore==1.37.11
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-nested-admin==4.1.1
django-storages==1.14.5
djangorestframework==3.15.2
environ==1.0
frozenlist==1.5.0
idna==3.10
jmespath==1.0.1
multidict==6.3.2
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
yarl==1.19.0
